requirements.txt
Flask==2.3.3
flask-cors==3.0.10
openai
bs4
python-dotenv
requests
numpy
mysql-connector-python
